requirements/runtime.txt
autotorch
matplotlib
mmpycocotools
numpy<2
opencv-python-headless
six
terminaltables
tqdm
